modify example
